Move environment reviews controllers out of artefacts module
